Do not verify just computed block fields
